% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_folds}
\alias{create_folds}
\title{Create Stratified K-Fold Splits}
\usage{
create_folds(ids, K)
}
\arguments{
\item{ids}{Vector of subject IDs.}

\item{K}{Integer number of folds.}
}
\value{
A list of length K, where each element contains the IDs assigned
to that fold.
}
\description{
Randomly assigns a vector of IDs into K approximately equal-sized folds.
}
\keyword{internal}
