% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_jfm.R
\name{jfm_counting_process}
\alias{jfm_counting_process}
\title{Counting Process of Recurrent Event}
\usage{
jfm_counting_process(t, recur_time)
}
\arguments{
\item{t}{A scalar time point.}

\item{recur_time}{A numeric vector of recurrent event times for a subject.}
}
\value{
Integer count of recurrent events at or before time t.
}
\description{
Computes the value of the counting process N_i(t), i.e., the number of
recurrent events that occurred at or before time t.
}
\keyword{internal}
