% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_jfm.R
\name{jfm_est_theta_new}
\alias{jfm_est_theta_new}
\title{Theta Estimating Equation for the Stagewise Algorithm}
\usage{
jfm_est_theta_new(
  theta,
  t.start,
  I,
  Z,
  alpha,
  beta,
  lambda0_r,
  lambda0_d,
  td,
  tr,
  tr.id,
  td.id,
  Y,
  STATUS,
  list_recur,
  num_recur
)
}
\arguments{
\item{theta}{Frailty variance parameter.}

\item{t.start}{Vector of interval start times.}

\item{I}{Vector of subject indicators for each pseudo entry.}

\item{Z}{List of covariate matrices, one per subject.}

\item{alpha}{Coefficient vector for the death sub-model.}

\item{beta}{Coefficient vector for the recurrent event sub-model.}

\item{lambda0_r}{Baseline hazard point masses for recurrence.}

\item{lambda0_d}{Baseline hazard point masses for death.}

\item{td}{Vector of death event times.}

\item{tr}{Vector of recurrent event times.}

\item{tr.id}{Subject IDs corresponding to each recurrent event.}

\item{td.id}{Subject IDs corresponding to each death event.}

\item{Y}{Vector of composite censoring/death times per subject.}

\item{STATUS}{Vector of death indicators per subject.}

\item{list_recur}{List of recurrent event times per subject.}

\item{num_recur}{Integer vector of recurrent event counts per subject.}
}
\value{
Scalar value of the scaled estimating equation for theta.
}
\description{
Self-contained version of the theta estimating equation that rebuilds
all intermediate quantities (wt_matrix, r2i, G-bar) internally, used
as the objective in a root-finding procedure for theta.
}
\keyword{internal}
