% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_jfm.R
\name{jfm_exp_icdf}
\alias{jfm_exp_icdf}
\title{Inverse CDF of Exponential Distribution}
\usage{
jfm_exp_icdf(y, rate)
}
\arguments{
\item{y}{Probability value(s) in (0, 1).}

\item{rate}{Rate parameter of the exponential distribution.}
}
\value{
Numeric value(s) corresponding to the inverse CDF.
}
\description{
Computes the quantile function (inverse CDF) of an exponential
distribution for simulation purposes.
}
\keyword{internal}
