% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_jfm.R
\name{jfm_lambda0d_solution}
\alias{jfm_lambda0d_solution}
\title{Solve for Baseline Death Hazard lambda0_d}
\usage{
jfm_lambda0d_solution(td, d_td, n, S0t_de)
}
\arguments{
\item{td}{Vector of death event times.}

\item{d_td}{Table of death time frequencies.}

\item{n}{Number of subjects.}

\item{S0t_de}{S0(t) vector for death.}
}
\value{
Named numeric vector of baseline hazard point masses for death.
}
\description{
Computes the closed-form solution for the baseline hazard point masses
of the death sub-model.
}
\keyword{internal}
