% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_jfm.R
\name{jfm_lambda0r_solution}
\alias{jfm_lambda0r_solution}
\title{Solve for Baseline Recurrence Hazard lambda0_r}
\usage{
jfm_lambda0r_solution(tr, d_tr, n, S0t_re)
}
\arguments{
\item{tr}{Vector of recurrent event times.}

\item{d_tr}{Table of recurrent event frequencies.}

\item{n}{Number of subjects.}

\item{S0t_re}{S0(t) vector for recurrence.}
}
\value{
Named numeric vector of baseline hazard point masses for recurrence.
}
\description{
Computes the closed-form solution for the baseline hazard point masses
of the recurrence sub-model.
}
\keyword{internal}
