% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_jfm.R
\name{jfm_r2i_integral}
\alias{jfm_r2i_integral}
\title{Integral in R2i(t) Evaluated at Each Recurrent Event Time}
\usage{
jfm_r2i_integral(t.start, I, Z, beta, tr, lambda0_r, tr.id)
}
\arguments{
\item{t.start}{Vector of interval start times.}

\item{I}{Vector of subject indicators for each pseudo entry.}

\item{Z}{List of covariate matrices, one per subject.}

\item{beta}{Coefficient vector for the recurrent event sub-model.}

\item{tr}{Vector of recurrent event times.}

\item{lambda0_r}{Baseline hazard point masses for recurrence.}

\item{tr.id}{Subject IDs corresponding to each recurrent event.}
}
\value{
A list with components:
\item{integral_matrix}{Matrix of integral values (rows = subjects, cols = recurrent times).}
\item{index_recurrent_matrix}{Matrix of pseudo entry indices.}
\item{tr_id}{Reordered subject IDs for recurrent events.}
}
\description{
Computes the integral component of R2i(t) for each subject at each
recurrent event time, using the pseudo data set approach.
}
\keyword{internal}
