% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_jfm.R
\name{jfm_s1t_recurrent}
\alias{jfm_s1t_recurrent}
\title{S1(t) for Recurrent Event Sub-Model}
\usage{
jfm_s1t_recurrent(
  Y,
  wt_recurrent_subject,
  tr,
  index_recurrent_matrix,
  pseudo_entries,
  beta
)
}
\arguments{
\item{Y}{Vector of composite censoring/death times per subject.}

\item{wt_recurrent_subject}{Weight matrix at recurrent event times.}

\item{tr}{Vector of recurrent event times.}

\item{index_recurrent_matrix}{Matrix of pseudo entry indices for recurrent times.}

\item{pseudo_entries}{Sorted pseudo data set.}

\item{beta}{Coefficient vector for the recurrent event sub-model.}
}
\value{
Matrix of S1(t) values (rows = covariates, cols = recurrent times).
}
\description{
Computes the weighted first moment S1(t) at each recurrent event time
for the recurrence sub-model.
}
\keyword{internal}
