% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_jfm.R
\name{jfm_score_alpha}
\alias{jfm_score_alpha}
\title{Score Equation U2 for Alpha (Death)}
\usage{
jfm_score_alpha(pseudo_entries, index_death_matrix, td_id, S1t_de, S0t_de)
}
\arguments{
\item{pseudo_entries}{Sorted pseudo data set.}

\item{index_death_matrix}{Matrix of pseudo entry indices for death times.}

\item{td_id}{Reordered subject IDs for death events.}

\item{S1t_de}{S1(t) matrix for death.}

\item{S0t_de}{S0(t) vector for death.}
}
\value{
Numeric vector of score values, one per covariate.
}
\description{
Evaluates the score equation for the death coefficient vector alpha.
}
\keyword{internal}
