% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_jfm.R
\name{jfm_v_gene}
\alias{jfm_v_gene}
\title{Generate a Single Covariate Vector}
\usage{
jfm_v_gene(n, v.type, v.coeffi)
}
\arguments{
\item{n}{Number of observations.}

\item{v.type}{Distribution type: 1 = normal, 2 = Bernoulli, 3 = Poisson.}

\item{v.coeffi}{Numeric vector of distribution parameters. For normal:
c(mean, sd). For Bernoulli: c(prob). For Poisson: c(lambda).}
}
\value{
Numeric vector of length n.
}
\description{
Generates n observations from a specified distribution for simulation
purposes.
}
\keyword{internal}
