% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_jfm.R
\name{jfm_wt_recurrent}
\alias{jfm_wt_recurrent}
\title{W_i(t) Evaluated at Each Recurrent Event Time for Each Subject}
\usage{
jfm_wt_recurrent(tr, wt_matrix, td)
}
\arguments{
\item{tr}{Vector of recurrent event times.}

\item{wt_matrix}{Weight matrix from \code{jfm_wt_death}.}

\item{td}{Vector of death event times.}
}
\value{
Matrix of W_i(t) values at recurrent event times
(rows = subjects, cols = recurrent event times).
}
\description{
Constructs the weight matrix W_i(t) evaluated at each recurrent event time
for each subject, by mapping recurrent event times to the nearest death
time in the weight matrix.
}
\keyword{internal}
