% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_jfm.R
\name{jfm_yt}
\alias{jfm_yt}
\title{At-Risk Indicator Y(t)}
\usage{
jfm_yt(y, t)
}
\arguments{
\item{y}{The subject's observed time.}

\item{t}{A scalar time point.}
}
\value{
Integer 0 or 1 indicator.
}
\description{
Returns 1 if the subject's observed time y is at least t, 0 otherwise.
}
\keyword{internal}
